Move test to test_utils.
